refactor: 플레이어 프로퍼티의 BetAmount를 제거 및 BetAmount 클래스 패키지 수정
